Broke down calver module to sub modules for clearer picture of dependencies. See #14
