+++ bumpit/core/versions/calver/constants.py
@@ -1,0 +1,10 @@
+YEAR_PART = "year"
+MONTH_PART = "month"
+DAY_PART = "day"
+MAJOR_PART = "major"
+MINOR_PART = "minor"
+MICRO_PART = "micro"
+MODIFIER_PART = "modifier"
+
+
+DATE_PARTS = [YEAR_PART, MONTH_PART, DAY_PART]
